refactor: move netutil to internal package and standardize naming
- Relocate netutil functions from `cmd/identity/bastion` to `internal/services/util` for improved code organization.
- Rename exported functions with PascalCase for consistency (`ExtractHostname`, `ResolveHostToIP`, `IsLocalTCPPortInUse`).
- Adjust references to renamed functions across Bastion-related flows.
- Remove unused SSH configuration logic and cleanup imports in bastion types.
- Fix typo in README ("releases" → "release").
