+++ internal/services/identity/bastion/types.go
@@ -1,13 +1,9 @@
 package bastion
 
 import (
-	"path/filepath"
-
 	"github.com/go-logr/logr"
 	"github.com/oracle/oci-go-sdk/v65/bastion"
 	"github.com/oracle/oci-go-sdk/v65/core"
-	conf "github.com/rozdolsky33/ocloud/internal/config"
-	cflags "github.com/rozdolsky33/ocloud/internal/config/flags"
 )
 
 // Bastion represents a bastion host in the system
@@ -37,27 +33,4 @@ type Config struct {
 	SshPublicKeyFile  string `yaml:"ssh_pub_key_file"`
 	SshPrivateKeyFile string `yaml:"ssh_private_key_file"`
 	SessionTimeout    *int   `yaml:"session_timeout"`
-}
-
-// sshConfig populates default values for SSH key locations based on the active OCI profile.
-// It dynamically constructs the path to the OCI session directory using standard constants.
-// The generated default points to the per-profile OCI session directory under:
-//
-//	<home>/.oci/sessions/<PROFILE>
-//
-// By default, it selects the standard OCI key filenames inside that directory:
-//   - oci_api_key_public.pem (public)
-//   - oci_api_key.pem (private)
-func (config *Config) sshConfig() {
-	homeDir, _ := conf.GetUserHomeDir()
-	profile := conf.GetOCIProfile()
-	// Build session directory: ~/.oci/sessions/<profile>
-	sessionDir := filepath.Join(homeDir, cflags.OCIConfigDirName, cflags.OCISessionsDirName, profile)
-
-	if config.SshPublicKeyFile == "" {
-		config.SshPublicKeyFile = filepath.Join(sessionDir, "oci_api_key_public.pem")
-	}
-	if config.SshPrivateKeyFile == "" {
-		config.SshPrivateKeyFile = filepath.Join(sessionDir, "oci_api_key.pem")
-	}
 }
